requirements/common.txt
# requirements/common.txt

Django==1.6.2

# Configuring databases
dj-database-url==0.2.2

# Managing static assets
dj-static==0.0.5
django-compressor==1.3

# Making navbar menus easier
django-lineage==0.2.0

# Customized to not always print out the json module being used
-e git+https://github.com/michaelbrooks/django-jsonview.git@no-logging#egg=django-jsonview

# Interfacing with Twitter
tweepy==2.2
-e git+https://github.com/michaelbrooks/twitter-monitor.git@simpler-threading#egg=twitter-monitor

django-bootstrap3==2.5.6
django-jsonview==0.3.0

nltk==2.0.4